Make edit filenames relative to the active window cwd by default
This lets a window be in some rando dir, use fzf, and have it open in
vis in another window and cwd. Don't use relativity for history as
it's already relative to the editor. Have hint_rg do absolutizing
itself since to let us use it from other windows eventually.